This refactors the fakerp package a bit
- admin_handlers.go now contains admin-related handlers
- customer_handlers.go now contains unprivileged handlers (basic createorupdate, get etc)
- routes.go now contains the routes specification and should be easier to deal with

/cc @mjudeikis @kwoodson @thekad
